# Server/api/openapi.yaml
openapi: 3.0.0
info:
  description: This is KNS Project API service.
  version: 1.0.2
  title: KNS - Incentive Community Project
  contact:
    email: rlawkdgks11@gmail.com
servers:
  - url: http://localhost:4000
...
paths:
...
  /post:
    post:
      tags:
        - post
      summary: 게시글 작성
      description: |
        게시글 작성
      operationId: writePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRegistration'
        required: true
      responses:
        '201':
          description: successful created
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '401':
          description: 인가되지 않은 사용자입니다.
  
    get:
      tags:
        - post
      summary: 전체 게시글 가져오기
      description: |
        등록한 전체 게시글 정보를 가져온다.
      operationId: getAllPosts
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostAll'
  /post/{postId}:
    put:
      tags:
        - post
      summary: 게시글 수정
      description: |
        게시글 수정
      operationId: updatePost
      parameters:
        - name: postId
          in: path
          required: true
          description: 게시글 아이디
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '403':
          description: 올바른 사용자가 아닙니다.
        '404':
          description: 글번호 몇 번이 존재하지 않습니다.

    delete:
      tags:
        - post
      summary: 게시글 삭제
      description: |
        게시글 삭제
      operationId: deletePost
      parameters:
        - name: postId
          in: path
          required: true
          description: 게시글 아이디
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '403':
          description: 올바른 사용자가 아닙니다.
        '404':
          description: 글번호 몇 번이 존재하지 않습니다.

    get:
      tags:
        - post
      summary: Post ID를 받아서 해당 게시글 상세 정보 응답
      description: |
        게시글 상세 정보 가져오기
      operationId: getPostById
      parameters:
        - name: postId
          in: path
          required: true
          description: 게시글 아이디
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostApiResponse'
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '404':
          description: 글번호 몇 번이 존재하지 않습니다.
  /post/like/{postId}:
    put:
      tags:
        - post
      summary: 게시글 좋아요 업데이트
      description: |
        게시글 좋아요 업데이트
      operationId: likePost
      parameters:
        - name: postId
          in: path
          required: true
          description: 게시글 아이디
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '404':
          description: 글번호 몇 번이 존재하지 않습니다.
  /post/{postId}/comment:
    post:
      tags:
        - comment
      summary: 댓글 작성
      description: |
        댓글 작성
      operationId: writeComment
      parameters:
        - name: postId
          in: path
          required: true
          description: 게시글 아이디
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRegistration'
        required: true
      responses:
        '201':
          description: successful created
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '404':
          description: 글번호 몇 번이 존재하지 않습니다.
  /post/comment/{commentId}:
    put:
      tags:
        - comment
      summary: 댓글 수정
      description: |
        댓글 수정
      operationId: updateComment
      parameters:
        - name: commentId
          in: path
          required: true
          description: 댓글 아이디
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '403':
          description: 올바른 사용자가 아닙니다.

    delete:
      tags:
        - comment
      summary: 댓글 삭제
      description: |
        댓글 삭제
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          required: true
          description: 댓글 아이디
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '403':
          description: 올바른 사용자가 아닙니다.

    get:
      tags:
        - comment
      summary: 댓글 정보 가져오기
      description: |
        댓글 정보 가져오기
      operationId: getCommentById
      parameters:
        - name: commentId
          in: path
          required: true
          description: 댓글 아이디
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '404':
          description: 댓글이 없습니다.
...
  /token/{userId}/transfer:
    put:
      tags:
        - token
      summary: ERC20 토큰 전송
      description: |
        ERC20 토큰 전송
      operationId: transferToken
      parameters:
        - name: userId
          in: path
          required: true
          description: 토큰 전송할 User ID
          schema:
            type: string
        - name: toUserId
          in: query
          description: 토큰 전송 대상 User ID
          schema:
            type: string
        - name: toTokenAmount
          in: query
          description: 전송할 토큰 수량
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
        '400':
          description: 올바르지 않은 파라미터 값입니다.
        '404':
          description: 토큰 전송할 사용자 정보가 없습니다.
        '409':
          description: 보유 토큰보다 전송할 토큰이 더 많습니다.
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        title:
          type: string
        content:
          type: string
        img:
          type: string
        hit:
          type: integer
          format: int32
        like:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
          description: create date
          example: '2021-01-30T08:30:20.000Z'
        updatedAt:
          type: string
          format: date-time
          description: update date
          example: '2021-01-30T08:30:20.000Z'
      required:
        - id
      example: {
                  "status": "successful operation",
                  "data": {
                      "id": 3,
                      "userId": 1,
                      "userName": "dhkim",
                      "title": "kdh",
                      "content": "12345",
                      "img": "",
                      "hit": 0,
                      "like": 0,
                      "createdAt": "2022-08-22T09:57:58.000Z",
                      "postComments": []
                  }
              }
    PostRegistration:
      type: object
      title: Post registration details
      properties:
        userId:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        img:
          type: string
      required:
        - userId
        - title
        - content
      example: null
    PostUpdate:
      type: object
      title: Post update details
      properties:
        title:
          type: string
        content:
          type: string
        img:
          type: string
      required:
        - title
        - content
      example: null
    PostAll:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        title:
          type: string
        img:
          type: string
        hit:
          type: integer
          format: int32
        like:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
          description: Creation date
          example: '2021-01-30T08:30:20.000Z'
        commentsCount:
          type: integer
          format: int32
      example: null
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
          description: Creation date
          example: '2021-01-30T08:30:20.000Z'
      required:
        - id
        - postId
        - userId
        - userName
        - comment
      example: null
    CommentRegistration:
      type: object
      title: Comment registration details
      properties:
        postId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        comment:
          type: string
      required:
        - postId
        - userId
        - userName
        - content
      example: null
    CommentUpdate:
      type: object
      title: Comment update details
      properties:
        comment:
          type: string
      required:
        - comment
      example: null
...
    PostApiResponse:
      type: array
      items:
        $ref: '#/components/schemas/Post'
      example: null
...